Disable team management from v7 on
